Make server side build precompile scripts rather than doing it at runtime (client build not working yet)
