+++ plugins/default/typescript/compiler/compileGame.ts
@@ -1,0 +1,84 @@
+/// <reference path="../typescriptAPI/TypeScriptAPIPlugin.d.ts" />
+
+import compileTypescript from "./compileTypeScript";
+import * as convert from "convert-source-map";
+
+// No definition file for combine-source-map module
+/* tslint:disable-next-line */
+const combine: any = require("combine-source-map");
+
+export default function compileGame(server: ProjectServer, scriptNames: string[], scripts: { [name: string]: string }, callback: (err: string, code: string) => void) {
+  const globalNames: string[] = [];
+  const globals: {[name: string]: string} = {};
+  const globalDefs: {[name: string]: string} = {};
+
+  const actorComponentAccessors: string[] = [];
+
+  const plugins = server.system.getPlugins<SupCore.TypeScriptAPIPlugin>("typescriptAPI");
+  for (const pluginName in plugins) {
+    const plugin = plugins[pluginName];
+    if (plugin.code != null) {
+      globalNames.push(`${pluginName}.ts`);
+      globals[`${pluginName}.ts`] = plugin.code;
+    }
+
+    if (plugin.defs != null) globalDefs[`${pluginName}.d.ts`] = plugin.defs;
+    if (plugin.exposeActorComponent != null) actorComponentAccessors.push(`${plugin.exposeActorComponent.propertyName}: ${plugin.exposeActorComponent.className};`);
+  }
+
+  // Plug component accessors exposed by plugins into Sup.Actor class
+  const joinedActorComponentAccessors = actorComponentAccessors.join("\n    ");
+  globals["Sup.Actor.ts"] = globals["Sup.Actor.ts"].replace("// INSERT_COMPONENT_ACCESSORS", joinedActorComponentAccessors);
+  globalDefs["Sup.Actor.d.ts"] = globalDefs["Sup.Actor.d.ts"].replace("// INSERT_COMPONENT_ACCESSORS", joinedActorComponentAccessors);
+
+  // Make sure the Sup namespace, Sup.Actor and Sup.ActorComponent are compiled before everything else
+  globalNames.unshift(globalNames.splice(globalNames.indexOf("Sup.Actor.ts"), 1)[0]);
+  globalNames.unshift(globalNames.splice(globalNames.indexOf("Sup.ts"), 1)[0]);
+
+  // Compile plugin globals
+  const libSource = `${globalDefs["lib.d.ts"]}\ndeclare var console, window, localStorage, player, SupEngine, SupRuntime, require;`;
+  const jsGlobals = compileTypescript(globalNames, globals, libSource, { sourceMap: false });
+  if (jsGlobals.errors.length > 0) {
+    for (const error of jsGlobals.errors) console.log(`${error.file}(${error.position.line}): ${error.message}`);
+    callback("Compilation failed. Check the devtools (F12) for errors.", null);
+    return;
+  }
+
+  // Compile game scripts
+  let concatenatedGlobalDefs = "";
+  for (const name in globalDefs) concatenatedGlobalDefs += globalDefs[name];
+  const results = compileTypescript(scriptNames, scripts, concatenatedGlobalDefs, { sourceMap: true });
+  if (results.errors.length > 0) {
+    for (const error of results.errors) console.log(`${error.file}(${error.position.line}): ${error.message}`);
+    callback("Compilation failed. Check the devtools (F12) for errors.", null);
+    return;
+  }
+
+  // Prepare source maps
+  const getLineCounts = (text: string) => {
+    let count = 1, index = -1;
+    while (true) {
+      index = text.indexOf("\n", index + 1);
+      if (index === -1) break;
+      count++;
+    }
+    return count;
+  };
+
+  jsGlobals.script = `(function() {
+var player = _player; _player = undefined;
+${jsGlobals.script}
+})();
+`;
+
+  let line = getLineCounts(jsGlobals.script) + 2;
+  const combinedSourceMap = combine.create("bundle.js");
+  for (const file of results.files) {
+    const comment = convert.fromObject(results.sourceMaps[file.name]).toComment();
+    combinedSourceMap.addFile({ sourceFile: `${server.data.manifest.pub.name}/${file.name}`, source: file.text + `\n${comment}` }, { line });
+    line += getLineCounts(file.text);
+  }
+
+  const code = `${jsGlobals.script}${results.script}//# sourceMappingURL=data:application/json;charset=utf-8;base64,${combinedSourceMap.base64()}`;
+  callback(null, code);
+}
